fix(server): satisfy test placement contract
Co-locate the Slice B tests with focused production modules so the server test-home inventory can keep shrinking. Preserve the existing contract, executor, integration, and opt-in dogfood coverage.
