+++ crates/harness-server/src/workflow_runtime_worker/mod.rs
@@ -3,6 +3,7 @@ mod activity_result;
 mod activity_status_contract;
 mod agent_contract_attempt;
 mod agent_contract_enforcement;
+mod agent_contract_job;
 mod child_workflow;
 mod child_workflow_non_issue;
 mod child_workflow_replay;
